Only need to fix mpu bug

+++ src/servos.h
@@ -3,7 +3,7 @@
 
 void handle_servo_change();
 void setup_servos();
-void set_throttle(float throttle);
+void set_throttle(long throttle);
 void set_pitch(long displacement);
 void set_roll(long displacement);
